feat: modified project structure for cucumber
