--- configs/chrome.mocha.conf.js
@@ -1,21 +1,0 @@
-import { downloadDir, mainConfig } from "../framework/configs/main.wdio.conf.js";
-
-
-export const config = {
-    ...mainConfig,
-    ...{
-        specs: [
-            '../test/specs/**/*.js'
-        ],
-        capabilities: [
-            {
-                browserName: "chrome",
-                "goog:chromeOptions": {
-                    prefs: {
-                        "download.default_directory": downloadDir
-                    }
-                },
-            },
-        ],
-    },
-};
