+++ package.json
@@ -2,7 +2,7 @@
   "name": "a1qa-wdio-framework",
   "type": "module",
   "scripts": {
-    "chrome.mocha": "wdio run ./configs/chrome.mocha.conf.js"
+    "chrome.cucumber": "wdio run ./configs/chrome.cucumber.conf.js"
   },
   "dependencies": {
     "@babel/core": "^7.24.3",
@@ -10,8 +10,8 @@
     "@babel/register": "^7.23.7",
     "@wdio/allure-reporter": "^8.34.0",
     "@wdio/cli": "^8.35.1",
+    "@wdio/cucumber-framework": "^8.36.1",
     "@wdio/local-runner": "^8.35.1",
-    "@wdio/mocha-framework": "^8.35.0",
     "@wdio/spec-reporter": "^8.32.4",
     "chai": "^5.1.0",
     "fs-extra": "^11.2.0",
